Refactor index API into resources
