+++ quickwit/quickwit-serve/src/index_api/mod.rs
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod index_resource;
 mod rest_handler;
+mod source_resource;
+mod split_resource;
 
-pub use self::rest_handler::{
-    get_index_metadata_handler, index_management_handlers, IndexApi, ListSplitsQueryParams,
-    ListSplitsResponse,
-};
+pub use self::index_resource::get_index_metadata_handler;
+pub use self::rest_handler::{index_management_handlers, IndexApi};
+pub use self::split_resource::{ListSplitsQueryParams, ListSplitsResponse};
