+++ quickwit/quickwit-serve/src/index_api/rest_handler.rs
@@ -14,39 +14,36 @@
 
 use std::sync::Arc;
 
-use bytes::Bytes;
-use quickwit_common::uri::Uri;
-use quickwit_config::{
-    load_index_config_update, load_source_config_from_user_config, validate_index_id_pattern,
-    ConfigFormat, FileSourceParams, NodeConfig, SourceConfig, SourceParams, CLI_SOURCE_ID,
-    INGEST_API_SOURCE_ID,
-};
+use quickwit_config::NodeConfig;
 use quickwit_doc_mapper::{analyze_text, TokenizerConfig};
 use quickwit_index_management::{IndexService, IndexServiceError};
-use quickwit_metastore::{
-    IndexMetadata, IndexMetadataResponseExt, ListIndexesMetadataResponseExt, ListSplitsQuery,
-    ListSplitsRequestExt, MetastoreServiceStreamSplitsExt, Split, SplitInfo, SplitState,
-    UpdateIndexRequestExt,
-};
-use quickwit_proto::ingest::Shard;
-use quickwit_proto::metastore::{
-    DeleteSourceRequest, EntityKind, IndexMetadataRequest, ListIndexesMetadataRequest,
-    ListShardsRequest, ListShardsSubrequest, ListSplitsRequest, MarkSplitsForDeletionRequest,
-    MetastoreError, MetastoreResult, MetastoreService, MetastoreServiceClient,
-    ResetSourceCheckpointRequest, ToggleSourceRequest, UpdateIndexRequest,
-};
-use quickwit_proto::types::{IndexId, IndexUid, SourceId};
 use quickwit_query::query_ast::{query_ast_from_user_text, QueryAst};
 use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
-use tracing::{info, warn};
+use serde::Deserialize;
+use tracing::warn;
 use warp::{Filter, Rejection};
 
-use crate::format::{extract_config_format, extract_format_from_qs};
+use super::get_index_metadata_handler;
+use super::index_resource::{
+    __path_clear_index, __path_create_index, __path_delete_index, __path_list_indexes_metadata,
+    __path_update_index, clear_index_handler, create_index_handler, delete_index_handler,
+    describe_index_handler, list_indexes_metadata_handler, update_index_handler, IndexStats,
+    __path_describe_index,
+};
+use super::source_resource::{
+    __path_create_source, __path_delete_source, __path_reset_source_checkpoint,
+    __path_toggle_source, create_source_handler, delete_source_handler, get_source_handler,
+    get_source_shards_handler, reset_source_checkpoint_handler, toggle_source_handler,
+    ToggleSource,
+};
+use super::split_resource::{
+    __path_list_splits, __path_mark_splits_for_deletion, list_splits_handler,
+    mark_splits_for_deletion_handler, SplitsForDeletion,
+};
+use crate::format::extract_format_from_qs;
 use crate::rest::recover_fn;
 use crate::rest_api_response::into_rest_api_response;
-use crate::simple_list::{from_simple_list, to_simple_list};
-use crate::with_arg;
+use crate::simple_list::from_simple_list;
 
 #[derive(utoipa::OpenApi)]
 #[openapi(
@@ -68,7 +65,7 @@ use crate::with_arg;
 )]
 pub struct IndexApi;
 
-fn log_failure<T, E: std::fmt::Display>(
+pub fn log_failure<T, E: std::fmt::Display>(
     message: &'static str,
 ) -> impl Fn(Result<T, E>) -> Result<T, E> + Clone {
     move |result| {
@@ -77,6 +74,11 @@ fn log_failure<T, E: std::fmt::Display>(
         };
         result
     }
+}
+
+pub fn json_body<T: DeserializeOwned + Send>(
+) -> impl Filter<Extract = (T,), Error = warp::Rejection> + Clone {
+    warp::body::content_length_limit(1024 * 1024).and(warp::body::json())
 }
 
 pub fn index_management_handlers(
@@ -110,869 +112,6 @@ pub fn index_management_handlers(
         .or(parse_query_request_handler())
         .recover(recover_fn)
         .boxed()
-}
-
-fn json_body<T: DeserializeOwned + Send>(
-) -> impl Filter<Extract = (T,), Error = warp::Rejection> + Clone {
-    warp::body::content_length_limit(1024 * 1024).and(warp::body::json())
-}
-
-pub fn get_index_metadata_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String)
-        .and(warp::get())
-        .and(with_arg(metastore))
-        .then(get_index_metadata)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-async fn get_index_metadata(
-    index_id: IndexId,
-    metastore: MetastoreServiceClient,
-) -> MetastoreResult<IndexMetadata> {
-    info!(index_id = %index_id, "get-index-metadata");
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let index_metadata = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?;
-    Ok(index_metadata)
-}
-
-/// This struct represents the QueryString passed to
-/// the rest API to filter indexes.
-#[derive(Debug, Clone, Deserialize, Serialize, utoipa::IntoParams, utoipa::ToSchema, Default)]
-#[into_params(parameter_in = Query)]
-pub struct ListIndexesQueryParams {
-    #[serde(deserialize_with = "from_simple_list")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub index_id_patterns: Option<Vec<String>>,
-}
-
-fn list_indexes_metadata_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes")
-        .and(warp::get())
-        .and(serde_qs::warp::query(serde_qs::Config::default()))
-        .and(with_arg(metastore))
-        .then(list_indexes_metadata)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-/// Describes an index with its main information and statistics.
-#[derive(Serialize, Deserialize, utoipa::ToSchema)]
-struct IndexStats {
-    #[schema(value_type = String)]
-    pub index_id: IndexId,
-    #[schema(value_type = String)]
-    pub index_uri: Uri,
-    pub num_published_splits: usize,
-    pub size_published_splits: u64,
-    pub num_published_docs: u64,
-    pub size_published_docs_uncompressed: u64,
-    pub timestamp_field_name: Option<String>,
-    pub min_timestamp: Option<i64>,
-    pub max_timestamp: Option<i64>,
-}
-
-#[utoipa::path(
-    get,
-    tag = "Indexes",
-    path = "/indexes/{index_id}/describe",
-    responses(
-        (status = 200, description = "Successfully fetched stats about Index.", body = IndexStats)
-    ),
-    params(
-        ("index_id" = String, Path, description = "The index ID to describe."),
-    )
-)]
-
-/// Describes an index.
-async fn describe_index(
-    index_id: IndexId,
-    metastore: MetastoreServiceClient,
-) -> MetastoreResult<IndexStats> {
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let index_metadata = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?;
-    let query = ListSplitsQuery::for_index(index_metadata.index_uid.clone());
-    let list_splits_request = ListSplitsRequest::try_from_list_splits_query(&query)?;
-    let splits = metastore
-        .list_splits(list_splits_request)
-        .await?
-        .collect_splits()
-        .await?;
-    let published_splits: Vec<Split> = splits
-        .into_iter()
-        .filter(|split| split.split_state == SplitState::Published)
-        .collect();
-    let mut total_num_docs = 0;
-    let mut total_num_bytes = 0;
-    let mut total_uncompressed_num_bytes = 0;
-    let mut min_timestamp: Option<i64> = None;
-    let mut max_timestamp: Option<i64> = None;
-
-    for split in &published_splits {
-        total_num_docs += split.split_metadata.num_docs as u64;
-        total_num_bytes += split.split_metadata.footer_offsets.end;
-        total_uncompressed_num_bytes += split.split_metadata.uncompressed_docs_size_in_bytes;
-
-        if let Some(time_range) = &split.split_metadata.time_range {
-            min_timestamp = min_timestamp
-                .min(Some(*time_range.start()))
-                .or(Some(*time_range.start()));
-            max_timestamp = max_timestamp
-                .max(Some(*time_range.end()))
-                .or(Some(*time_range.end()));
-        }
-    }
-
-    let index_config = index_metadata.into_index_config();
-    let index_stats = IndexStats {
-        index_id,
-        index_uri: index_config.index_uri.clone(),
-        num_published_splits: published_splits.len(),
-        size_published_splits: total_num_bytes,
-        num_published_docs: total_num_docs,
-        size_published_docs_uncompressed: total_uncompressed_num_bytes,
-        timestamp_field_name: index_config.doc_mapping.timestamp_field,
-        min_timestamp,
-        max_timestamp,
-    };
-
-    Ok(index_stats)
-}
-
-fn describe_index_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "describe")
-        .and(warp::get())
-        .and(with_arg(metastore))
-        .then(describe_index)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-/// This struct represents the QueryString passed to
-/// the rest API to filter splits.
-#[derive(Debug, Clone, Deserialize, Serialize, utoipa::IntoParams, utoipa::ToSchema, Default)]
-#[into_params(parameter_in = Query)]
-pub struct ListSplitsQueryParams {
-    /// If set, define the number of splits to skip
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub offset: Option<usize>,
-    /// If set, restrict maximum number of splits to retrieve
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub limit: Option<usize>,
-    /// A specific split state(s) to filter by.
-    #[serde(deserialize_with = "from_simple_list")]
-    #[serde(serialize_with = "to_simple_list")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub split_states: Option<Vec<SplitState>>,
-    /// If set, restrict splits to documents with a `timestamp >= start_timestamp`.
-    /// This timestamp is in seconds.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub start_timestamp: Option<i64>,
-    /// If set, restrict splits to documents with a `timestamp < end_timestamp`.
-    /// This timestamp is in seconds.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub end_timestamp: Option<i64>,
-    /// If set, restrict splits whose creation dates are before this date.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[serde(default)]
-    pub end_create_timestamp: Option<i64>,
-}
-
-#[derive(Serialize, Deserialize, Debug, utoipa::ToSchema)]
-pub struct ListSplitsResponse {
-    #[serde(default)]
-    pub offset: usize,
-    #[serde(default)]
-    pub size: usize,
-    #[serde(default)]
-    pub splits: Vec<Split>,
-}
-
-#[utoipa::path(
-    get,
-    tag = "Indexes",
-    path = "/indexes/{index_id}/splits",
-    responses(
-        (status = 200, description = "Successfully fetched splits.", body = ListSplitsResponse)
-    ),
-    params(
-        ListSplitsQueryParams,
-        ("index_id" = String, Path, description = "The index ID to retrieve splits for."),
-    )
-)]
-
-/// Get splits.
-async fn list_splits(
-    index_id: IndexId,
-    list_split_query: ListSplitsQueryParams,
-    metastore: MetastoreServiceClient,
-) -> MetastoreResult<ListSplitsResponse> {
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let index_uid: IndexUid = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?
-        .index_uid;
-    info!(index_id = %index_id, list_split_query = ?list_split_query, "get-splits");
-    let mut query = ListSplitsQuery::for_index(index_uid);
-    let mut offset = 0;
-    if let Some(offset_value) = list_split_query.offset {
-        query = query.with_offset(offset_value);
-        offset = offset_value;
-    }
-    if let Some(limit) = list_split_query.limit {
-        query = query.with_limit(limit);
-    }
-    if let Some(split_states) = list_split_query.split_states {
-        query = query.with_split_states(split_states);
-    }
-    if let Some(start_timestamp) = list_split_query.start_timestamp {
-        query = query.with_time_range_start_gte(start_timestamp);
-    }
-    if let Some(end_timestamp) = list_split_query.end_timestamp {
-        query = query.with_time_range_end_lt(end_timestamp);
-    }
-    if let Some(end_created_timestamp) = list_split_query.end_create_timestamp {
-        query = query.with_create_timestamp_lt(end_created_timestamp);
-    }
-    let list_splits_request = ListSplitsRequest::try_from_list_splits_query(&query)?;
-    let splits = metastore
-        .list_splits(list_splits_request)
-        .await?
-        .collect_splits()
-        .await?;
-    Ok(ListSplitsResponse {
-        offset,
-        size: splits.len(),
-        splits,
-    })
-}
-
-fn list_splits_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "splits")
-        .and(warp::get())
-        .and(serde_qs::warp::query(serde_qs::Config::default()))
-        .and(with_arg(metastore))
-        .then(list_splits)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[derive(Deserialize, utoipa::ToSchema)]
-#[serde(deny_unknown_fields)]
-struct SplitsForDeletion {
-    pub split_ids: Vec<String>,
-}
-
-#[utoipa::path(
-    put,
-    tag = "Splits",
-    path = "/indexes/{index_id}/splits/mark-for-deletion",
-    request_body = SplitsForDeletion,
-    responses(
-        (status = 200, description = "Successfully marked splits for deletion.")
-    ),
-    params(
-        ("index_id" = String, Path, description = "The index ID to mark splits for deletion for."),
-    )
-)]
-/// Marks splits for deletion.
-async fn mark_splits_for_deletion(
-    index_id: IndexId,
-    splits_for_deletion: SplitsForDeletion,
-    metastore: MetastoreServiceClient,
-) -> MetastoreResult<()> {
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let index_uid: IndexUid = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?
-        .index_uid;
-    info!(index_id = %index_id, splits_ids = ?splits_for_deletion.split_ids, "mark-splits-for-deletion");
-    let split_ids: Vec<String> = splits_for_deletion
-        .split_ids
-        .iter()
-        .map(|split_id| split_id.to_string())
-        .collect();
-    let mark_splits_for_deletion_request =
-        MarkSplitsForDeletionRequest::new(index_uid, split_ids.clone());
-    metastore
-        .mark_splits_for_deletion(mark_splits_for_deletion_request)
-        .await?;
-    Ok(())
-}
-
-fn mark_splits_for_deletion_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "splits" / "mark-for-deletion")
-        .and(warp::put())
-        .and(json_body())
-        .and(with_arg(metastore))
-        .then(mark_splits_for_deletion)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[utoipa::path(
-    get,
-    tag = "Indexes",
-    path = "/indexes",
-    responses(
-        // We return `VersionedIndexMetadata` as it's the serialized model view.
-        (status = 200, description = "Successfully fetched all indexes.", body = [VersionedIndexMetadata])
-    ),
-    params(
-        ListIndexesQueryParams,
-        ("index_id_patterns" = String, Path, description = "The index ID pattern to retrieve indexes for."),
-    )
-)]
-/// Gets indexes metadata.
-async fn list_indexes_metadata(
-    list_indexes_params: ListIndexesQueryParams,
-    metastore: MetastoreServiceClient,
-) -> MetastoreResult<Vec<IndexMetadata>> {
-    let list_indexes_metata_request =
-        if let Some(index_id_patterns) = list_indexes_params.index_id_patterns {
-            for index_id_pattern in &index_id_patterns {
-                validate_index_id_pattern(index_id_pattern, true).map_err(|error| {
-                    MetastoreError::InvalidArgument {
-                        message: error.to_string(),
-                    }
-                })?;
-            }
-            ListIndexesMetadataRequest { index_id_patterns }
-        } else {
-            ListIndexesMetadataRequest::all()
-        };
-    metastore
-        .list_indexes_metadata(list_indexes_metata_request)
-        .await?
-        .deserialize_indexes_metadata()
-        .await
-}
-
-#[derive(Deserialize, utoipa::IntoParams, utoipa::ToSchema)]
-#[into_params(parameter_in = Query)]
-struct CreateIndexQueryParams {
-    #[serde(default)]
-    overwrite: bool,
-}
-
-fn create_index_handler(
-    index_service: IndexService,
-    node_config: Arc<NodeConfig>,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes")
-        .and(warp::post())
-        .and(serde_qs::warp::query(serde_qs::Config::default()))
-        .and(extract_config_format())
-        .and(warp::body::content_length_limit(1024 * 1024))
-        .and(warp::filters::body::bytes())
-        .and(with_arg(index_service))
-        .and(with_arg(node_config))
-        .then(create_index)
-        .map(log_failure("failed to create index"))
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[utoipa::path(
-    post,
-    tag = "Indexes",
-    path = "/indexes",
-    request_body = VersionedIndexConfig,
-    responses(
-        // We return `VersionedIndexMetadata` as it's the serialized model view.
-        (status = 200, description = "Successfully created index.", body = VersionedIndexMetadata)
-    ),
-    params(
-        CreateIndexQueryParams,
-    )
-)]
-/// Creates index.
-async fn create_index(
-    create_index_query_params: CreateIndexQueryParams,
-    config_format: ConfigFormat,
-    index_config_bytes: Bytes,
-    mut index_service: IndexService,
-    node_config: Arc<NodeConfig>,
-) -> Result<IndexMetadata, IndexServiceError> {
-    let index_config = quickwit_config::load_index_config_from_user_config(
-        config_format,
-        &index_config_bytes,
-        &node_config.default_index_root_uri,
-    )
-    .map_err(IndexServiceError::InvalidConfig)?;
-    info!(index_id = %index_config.index_id, overwrite = create_index_query_params.overwrite, "create-index");
-    index_service
-        .create_index(index_config, create_index_query_params.overwrite)
-        .await
-}
-
-fn update_index_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String)
-        .and(warp::put())
-        .and(extract_config_format())
-        .and(warp::body::content_length_limit(1024 * 1024))
-        .and(warp::filters::body::bytes())
-        .and(with_arg(metastore))
-        .then(update_index)
-        .map(log_failure("failed to update index"))
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[utoipa::path(
-    put,
-    tag = "Indexes",
-    path = "/indexes/{index_id}",
-    request_body = VersionedIndexConfig,
-    responses(
-        (status = 200, description = "Successfully updated the index configuration.", body = VersionedIndexMetadata)
-    ),
-    params(
-        ("index_id" = String, Path, description = "The index ID to update."),
-    )
-)]
-/// Updates an existing index.
-///
-/// This endpoint follows PUT semantics, which means that all the fields of the
-/// current configuration are replaced by the values specified in this request
-/// or the associated defaults. In particular, if the field is optional (e.g.
-/// `retention_policy`), omitting it will delete the associated configuration.
-/// If the new configuration file contains updates that cannot be applied, the
-/// request fails, and none of the updates are applied.
-async fn update_index(
-    target_index_id: IndexId,
-    config_format: ConfigFormat,
-    index_config_bytes: Bytes,
-    metastore: MetastoreServiceClient,
-) -> Result<IndexMetadata, IndexServiceError> {
-    info!(index_id = %target_index_id, "update-index");
-
-    let index_metadata_request = IndexMetadataRequest::for_index_id(target_index_id.to_string());
-    let current_index_metadata = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?;
-    let index_uid = current_index_metadata.index_uid.clone();
-    let current_index_config = current_index_metadata.into_index_config();
-
-    let new_index_config =
-        load_index_config_update(config_format, &index_config_bytes, &current_index_config)
-            .map_err(IndexServiceError::InvalidConfig)?;
-
-    let update_request = UpdateIndexRequest::try_from_updates(
-        index_uid,
-        &new_index_config.search_settings,
-        &new_index_config.retention_policy_opt,
-        &new_index_config.indexing_settings,
-        &new_index_config.doc_mapping,
-    )?;
-    let update_resp = metastore.update_index(update_request).await?;
-    Ok(update_resp.deserialize_index_metadata()?)
-}
-
-fn clear_index_handler(
-    index_service: IndexService,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "clear")
-        .and(warp::put())
-        .and(with_arg(index_service))
-        .then(clear_index)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[utoipa::path(
-    put,
-    tag = "Indexes",
-    path = "/indexes/{index_id}/clear",
-    responses(
-        (status = 200, description = "Successfully cleared index.")
-    ),
-    params(
-        ("index_id" = String, Path, description = "The index ID to clear."),
-    )
-)]
-/// Removes all of the data (splits, queued document) associated with the index, but keeps the index
-/// configuration. (See also, `delete-index`).
-async fn clear_index(
-    index_id: IndexId,
-    mut index_service: IndexService,
-) -> Result<(), IndexServiceError> {
-    info!(index_id = %index_id, "clear-index");
-    index_service.clear_index(&index_id).await
-}
-
-#[derive(Deserialize, utoipa::IntoParams, utoipa::ToSchema)]
-#[into_params(parameter_in = Query)]
-struct DeleteIndexQueryParam {
-    #[serde(default)]
-    dry_run: bool,
-}
-
-fn delete_index_handler(
-    index_service: IndexService,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String)
-        .and(warp::delete())
-        .and(serde_qs::warp::query(serde_qs::Config::default()))
-        .and(with_arg(index_service))
-        .then(delete_index)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[utoipa::path(
-    delete,
-    tag = "Indexes",
-    path = "/indexes/{index_id}",
-    responses(
-        // We return `VersionedIndexMetadata` as it's the serialized model view.
-        (status = 200, description = "Successfully deleted index.", body = [FileEntry])
-    ),
-    params(
-        DeleteIndexQueryParam,
-        ("index_id" = String, Path, description = "The index ID to delete."),
-    )
-)]
-/// Deletes index.
-async fn delete_index(
-    index_id: IndexId,
-    delete_index_query_param: DeleteIndexQueryParam,
-    mut index_service: IndexService,
-) -> Result<Vec<SplitInfo>, IndexServiceError> {
-    info!(index_id = %index_id, dry_run = delete_index_query_param.dry_run, "delete-index");
-    index_service
-        .delete_index(&index_id, delete_index_query_param.dry_run)
-        .await
-}
-
-fn create_source_handler(
-    index_service: IndexService,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "sources")
-        .and(warp::post())
-        .and(extract_config_format())
-        .and(warp::body::content_length_limit(1024 * 1024))
-        .and(warp::filters::body::bytes())
-        .and(with_arg(index_service))
-        .then(create_source)
-        .map(log_failure("failed to create source"))
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[utoipa::path(
-    post,
-    tag = "Sources",
-    path = "/indexes/{index_id}/sources",
-    request_body = VersionedSourceConfig,
-    responses(
-        // We return `VersionedSourceConfig` as it's the serialized model view.
-        (status = 200, description = "Successfully created source.", body = VersionedSourceConfig)
-    ),
-    params(
-        ("index_id" = String, Path, description = "The index ID to create a source for."),
-    )
-)]
-/// Creates Source.
-async fn create_source(
-    index_id: IndexId,
-    config_format: ConfigFormat,
-    source_config_bytes: Bytes,
-    mut index_service: IndexService,
-) -> Result<SourceConfig, IndexServiceError> {
-    let source_config: SourceConfig =
-        load_source_config_from_user_config(config_format, &source_config_bytes)
-            .map_err(IndexServiceError::InvalidConfig)?;
-    // Note: This check is performed here instead of the source config serde
-    // because many tests use the file source, and can't store that config in
-    // the metastore without going through the validation.
-    if let SourceParams::File(FileSourceParams::Filepath(_)) = &source_config.source_params {
-        return Err(IndexServiceError::InvalidConfig(anyhow::anyhow!(
-            "path based file sources are limited to a local usage, please use the CLI command \
-             `quickwit tool local-ingest` to ingest data from a specific file or setup a \
-             notification based file source"
-        )));
-    }
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let index_uid: IndexUid = index_service
-        .metastore()
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?
-        .index_uid;
-    info!(index_id = %index_id, source_id = %source_config.source_id, "create-source");
-    index_service.add_source(index_uid, source_config).await
-}
-
-fn get_source_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "sources" / String)
-        .and(warp::get())
-        .and(with_arg(metastore))
-        .then(get_source)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-async fn get_source(
-    index_id: IndexId,
-    source_id: SourceId,
-    metastore: MetastoreServiceClient,
-) -> MetastoreResult<SourceConfig> {
-    info!(index_id = %index_id, source_id = %source_id, "get-source");
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let source_config = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?
-        .sources
-        .remove(&source_id)
-        .ok_or({
-            MetastoreError::NotFound(EntityKind::Source {
-                index_id,
-                source_id,
-            })
-        })?;
-    Ok(source_config)
-}
-
-fn reset_source_checkpoint_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "sources" / String / "reset-checkpoint")
-        .and(warp::put())
-        .and(with_arg(metastore))
-        .then(reset_source_checkpoint)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[utoipa::path(
-    put,
-    tag = "Sources",
-    path = "/indexes/{index_id}/sources/{source_id}/reset-checkpoint",
-    responses(
-        (status = 200, description = "Successfully reset source checkpoint.")
-    ),
-    params(
-        ("index_id" = String, Path, description = "The index ID of the source."),
-        ("source_id" = String, Path, description = "The source ID whose checkpoint is reset."),
-    )
-)]
-/// Resets source checkpoint.
-async fn reset_source_checkpoint(
-    index_id: IndexId,
-    source_id: SourceId,
-    metastore: MetastoreServiceClient,
-) -> MetastoreResult<()> {
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let index_uid: IndexUid = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?
-        .index_uid;
-    info!(index_id = %index_id, source_id = %source_id, "reset-checkpoint");
-    let reset_source_checkpoint_request = ResetSourceCheckpointRequest {
-        index_uid: Some(index_uid),
-        source_id: source_id.clone(),
-    };
-    metastore
-        .reset_source_checkpoint(reset_source_checkpoint_request)
-        .await?;
-    Ok(())
-}
-
-fn toggle_source_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "sources" / String / "toggle")
-        .and(warp::put())
-        .and(json_body())
-        .and(with_arg(metastore))
-        .then(toggle_source)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[derive(Deserialize, utoipa::ToSchema)]
-#[serde(deny_unknown_fields)]
-struct ToggleSource {
-    enable: bool,
-}
-
-#[utoipa::path(
-    put,
-    tag = "Sources",
-    path = "/indexes/{index_id}/sources/{source_id}/toggle",
-    request_body = ToggleSource,
-    responses(
-        (status = 200, description = "Successfully toggled source.")
-    ),
-    params(
-        ("index_id" = String, Path, description = "The index ID of the source."),
-        ("source_id" = String, Path, description = "The source ID to toggle."),
-    )
-)]
-/// Toggles source.
-async fn toggle_source(
-    index_id: IndexId,
-    source_id: SourceId,
-    toggle_source: ToggleSource,
-    metastore: MetastoreServiceClient,
-) -> Result<(), IndexServiceError> {
-    info!(index_id = %index_id, source_id = %source_id, enable = toggle_source.enable, "toggle-source");
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let index_uid: IndexUid = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?
-        .index_uid;
-    if [CLI_SOURCE_ID, INGEST_API_SOURCE_ID].contains(&source_id.as_str()) {
-        return Err(IndexServiceError::OperationNotAllowed(format!(
-            "source `{source_id}` is managed by Quickwit, you cannot enable or disable a source \
-             managed by Quickwit"
-        )));
-    }
-    let toggle_source_request = ToggleSourceRequest {
-        index_uid: Some(index_uid),
-        source_id: source_id.clone(),
-        enable: toggle_source.enable,
-    };
-    metastore.toggle_source(toggle_source_request).await?;
-    Ok(())
-}
-
-fn delete_source_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "sources" / String)
-        .and(warp::delete())
-        .and(with_arg(metastore))
-        .then(delete_source)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-#[utoipa::path(
-    delete,
-    tag = "Sources",
-    path = "/indexes/{index_id}/sources/{source_id}",
-    responses(
-        (status = 200, description = "Successfully deleted source.")
-    ),
-    params(
-        ("index_id" = String, Path, description = "The index ID to remove the source from."),
-        ("source_id" = String, Path, description = "The source ID to remove from the index."),
-    )
-)]
-/// Deletes source.
-async fn delete_source(
-    index_id: IndexId,
-    source_id: SourceId,
-    metastore: MetastoreServiceClient,
-) -> Result<(), IndexServiceError> {
-    info!(index_id = %index_id, source_id = %source_id, "delete-source");
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let index_uid: IndexUid = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?
-        .index_uid;
-    if [INGEST_API_SOURCE_ID, CLI_SOURCE_ID].contains(&source_id.as_str()) {
-        return Err(IndexServiceError::OperationNotAllowed(format!(
-            "source `{source_id}` is managed by Quickwit, you cannot delete a source managed by \
-             Quickwit"
-        )));
-    }
-    let delete_source_request = DeleteSourceRequest {
-        index_uid: Some(index_uid),
-        source_id: source_id.clone(),
-    };
-    metastore.delete_source(delete_source_request).await?;
-    Ok(())
-}
-
-fn get_source_shards_handler(
-    metastore: MetastoreServiceClient,
-) -> impl Filter<Extract = (impl warp::Reply,), Error = Rejection> + Clone {
-    warp::path!("indexes" / String / "sources" / String / "shards")
-        .and(warp::get())
-        .and(with_arg(metastore))
-        .then(get_source_shards)
-        .and(extract_format_from_qs())
-        .map(into_rest_api_response)
-        .boxed()
-}
-
-async fn get_source_shards(
-    index_id: IndexId,
-    source_id: SourceId,
-    metastore: MetastoreServiceClient,
-) -> MetastoreResult<Vec<Shard>> {
-    info!(index_id = %index_id, source_id = %source_id, "get-source-shards");
-    let index_metadata_request = IndexMetadataRequest::for_index_id(index_id.to_string());
-    let index_uid: IndexUid = metastore
-        .index_metadata(index_metadata_request)
-        .await?
-        .deserialize_index_metadata()?
-        .index_uid;
-    let response = metastore
-        .list_shards(ListShardsRequest {
-            subrequests: vec![ListShardsSubrequest {
-                index_uid: Some(index_uid),
-                source_id: source_id.to_string(),
-                ..Default::default()
-            }],
-        })
-        .await?;
-    let shards = response
-        .subresponses
-        .into_iter()
-        .flat_map(|resp| resp.shards)
-        .collect();
-    Ok(shards)
 }
 
 #[derive(Debug, Deserialize, utoipa::IntoParams, utoipa::ToSchema)]
@@ -1069,13 +208,22 @@ mod tests {
     use assert_json_diff::assert_json_include;
     use quickwit_common::uri::Uri;
     use quickwit_common::ServiceStream;
-    use quickwit_config::{SourceParams, VecSourceParams};
-    use quickwit_indexing::{mock_split, MockSplitBuilder};
-    use quickwit_metastore::{metastore_for_test, IndexMetadata, ListSplitsResponseExt};
-    use quickwit_proto::metastore::{
-        EmptyResponse, IndexMetadataResponse, ListIndexesMetadataResponse, ListSplitsResponse,
-        MetastoreServiceClient, MockMetastoreService, SourceType,
+    use quickwit_config::{
+        NodeConfig, SourceParams, VecSourceParams, CLI_SOURCE_ID, INGEST_API_SOURCE_ID,
     };
+    use quickwit_indexing::{mock_split, MockSplitBuilder};
+    use quickwit_metastore::{
+        metastore_for_test, IndexMetadata, IndexMetadataResponseExt,
+        ListIndexesMetadataResponseExt, ListSplitsRequestExt, ListSplitsResponseExt, SplitState,
+    };
+    use quickwit_proto::metastore::{
+        DeleteSourceRequest, EmptyResponse, EntityKind, IndexMetadataRequest,
+        IndexMetadataResponse, ListIndexesMetadataRequest, ListIndexesMetadataResponse,
+        ListSplitsRequest, ListSplitsResponse, MarkSplitsForDeletionRequest, MetastoreError,
+        MetastoreService, MetastoreServiceClient, MockMetastoreService,
+        ResetSourceCheckpointRequest, SourceType, ToggleSourceRequest,
+    };
+    use quickwit_proto::types::IndexUid;
     use quickwit_storage::StorageResolver;
     use serde_json::Value as JsonValue;
